Create a new module to provide MockRetrofit instances and correct JSON value of platform_version.
