+++ data/src/test/java/org/oppia/android/data/backends/test/MockConceptCardTest.kt
@@ -2,19 +2,14 @@ package org.oppia.android.data.backends.test
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
-import okhttp3.OkHttpClient
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.oppia.android.data.backends.api.MockConceptCardService
-import org.oppia.android.data.backends.gae.NetworkInterceptor
-import org.oppia.android.data.backends.gae.NetworkSettings
 import org.oppia.android.data.backends.gae.api.ConceptCardService
+import org.oppia.android.testing.network.MockRetrofitHelper
 import org.robolectric.annotation.LooperMode
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.mock.MockRetrofit
-import retrofit2.mock.NetworkBehavior
 
 /**
  * Test for [ConceptCardService] retrofit instance using [MockConceptCardService]
@@ -23,24 +18,10 @@ import retrofit2.mock.NetworkBehavior
 @LooperMode(LooperMode.Mode.PAUSED)
 class MockConceptCardTest {
   private lateinit var mockRetrofit: MockRetrofit
-  private lateinit var retrofit: Retrofit
 
   @Before
   fun setUp() {
-    val client = OkHttpClient.Builder()
-    client.addInterceptor(NetworkInterceptor())
-
-    retrofit = Retrofit.Builder()
-      .baseUrl(NetworkSettings.getBaseUrl())
-      .addConverterFactory(MoshiConverterFactory.create())
-      .client(client.build())
-      .build()
-
-    val behavior = NetworkBehavior.create()
-    behavior.setFailurePercent(0)
-    mockRetrofit = MockRetrofit.Builder(retrofit)
-      .networkBehavior(behavior)
-      .build()
+    mockRetrofit = MockRetrofitHelper().createMockRetrofit()
   }
 
   @Test
