+++ data/src/test/java/org/oppia/android/data/backends/test/MockFeedbackReportingTest.kt
@@ -4,45 +4,26 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth.assertThat
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.Moshi
-import okhttp3.OkHttpClient
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.oppia.android.data.backends.ApiUtils
 import org.oppia.android.data.backends.api.MockFeedbackReportingService
-import org.oppia.android.data.backends.gae.NetworkInterceptor
-import org.oppia.android.data.backends.gae.NetworkSettings
 import org.oppia.android.data.backends.gae.api.FeedbackReportingService
 import org.oppia.android.data.backends.gae.model.GaeFeedbackReport
+import org.oppia.android.testing.network.MockRetrofitHelper
 import org.robolectric.annotation.LooperMode
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.mock.MockRetrofit
-import retrofit2.mock.NetworkBehavior
 
 /** Test for [FeedbackReportingService] retrofit instance using a [MockFeedbackReportingService]. */
 @RunWith(AndroidJUnit4::class)
 @LooperMode(LooperMode.Mode.PAUSED)
 class MockFeedbackReportingTest {
   private lateinit var mockRetrofit: MockRetrofit
-  private lateinit var retrofit: Retrofit
 
   @Before
   fun setUp() {
-    val client = OkHttpClient.Builder()
-    client.addInterceptor(NetworkInterceptor())
-
-    retrofit = Retrofit.Builder()
-      .baseUrl(NetworkSettings.getBaseUrl())
-      .addConverterFactory(MoshiConverterFactory.create())
-      .client(client.build())
-      .build()
-
-    val behavior = NetworkBehavior.create()
-    behavior.setFailurePercent(0)
-    mockRetrofit = MockRetrofit.Builder(retrofit)
-      .networkBehavior(behavior)
-      .build()
+    mockRetrofit = MockRetrofitHelper().createMockRetrofit()
   }
 
   @Test
